# docs/swagger.yaml
basePath: /
definitions:
  email-service_internal_model.Account:
    description: Holds credentials and configuration for an external email service.
    properties:
      access_key_id:
        type: string
      created_at:
        type: string
      daily_send_limit:
        type: integer
      domain:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  email-service_internal_model.AccountSender:
    description: Represents the many-to-many relationship between accounts and senders,
      with specific sending configurations.
    properties:
      account:
        $ref: '#/definitions/email-service_internal_model.Account'
      account_id:
        type: integer
      created_at:
        type: string
      daily_send_limit:
        type: integer
      email_address:
        type: string
      id:
        type: integer
      sender:
        $ref: '#/definitions/email-service_internal_model.Sender'
      sender_id:
        type: integer
      status:
        type: string
      updated_at:
        type: string
      weight:
        type: integer
    type: object
...
  email-service_internal_model.Recipient:
    description: Stores information about a person who can receive emails, including
      custom data.
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  email-service_internal_model.Sender:
    description: Contains information about the sender of an email.
    properties:
      contact_info:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
...
  internal_handler.AddSenderToAccountRequest:
    properties:
      daily_send_limit:
        example: 1000
        type: integer
      email_address:
        example: noreply@example.com
        type: string
    required:
    - daily_send_limit
    - email_address
    type: object
...
  internal_handler.CreateRecipientRequest:
    properties:
      email:
        example: john.doe@example.com
        type: string
      first_name:
        example: John
        type: string
      last_name:
        example: Doe
        type: string
      metadata:
        additionalProperties: true
        type: object
    required:
    - email
    type: object
...
  internal_handler.PaginatedResponse:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/internal_handler.Pagination'
    type: object
  internal_handler.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  internal_handler.Response:
    properties:
      data: {}
      error:
        type: string
    type: object
...
  internal_handler.UpdateRecipientRequest:
    properties:
      email:
        example: john.doe.updated@example.com
        type: string
      first_name:
        example: Johnathan
        type: string
      last_name:
        example: Doe
        type: string
      metadata:
        additionalProperties: true
        type: object
      status:
        example: unsubscribed
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: email-service/internal/handler
    url: http://www.swagger.io/support
  description: This is a powerful email service API for batch sending and management.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Email Service API
  version: "1.0"
paths:
...
  /api/v1/recipients:
    get:
      description: Retrieves a paginated list of all recipients.
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of items per page (default: 10)'
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/email-service_internal_model.Recipient'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: List Recipients
      tags:
      - Recipients
    post:
      consumes:
      - application/json
      description: Adds a new recipient to the system.
      parameters:
      - description: Recipient Information
        in: body
        name: recipient
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateRecipientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/email-service_internal_model.Recipient'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Create Recipient
      tags:
      - Recipients
  /api/v1/recipients/{id}:
    delete:
      description: Deletes a recipient from the system.
      parameters:
      - description: Recipient ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid recipient ID
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete Recipient
      tags:
      - Recipients
    get:
      description: Retrieves details for a specific recipient.
      parameters:
      - description: Recipient ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/email-service_internal_model.Recipient'
              type: object
        "400":
          description: Invalid recipient ID
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "404":
          description: Recipient not found
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Get Recipient by ID
      tags:
      - Recipients
    put:
      consumes:
      - application/json
      description: Updates the details of an existing recipient.
      parameters:
      - description: Recipient ID
        in: path
        name: id
        required: true
        type: integer
      - description: Recipient fields to update
        in: body
        name: recipient
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UpdateRecipientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/email-service_internal_model.Recipient'
              type: object
        "400":
          description: Invalid recipient ID or request body
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Update Recipient
      tags:
      - Recipients
  /api/v1/senders:
    post:
      consumes:
      - application/json
      description: Creates a new sender entity (e.g., a person or department).
      parameters:
      - description: Sender Information
        in: body
        name: sender
        required: true
        schema:
          $ref: '#/definitions/email-service_internal_model.Sender'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/email-service_internal_model.Sender'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Create Sender
      tags:
      - Senders
  /api/v1/senders/{senderId}/accounts/{accountId}:
    post:
      consumes:
      - application/json
      description: Associates a sender with an account, creating a usable 'from' email
        address with a specific sending limit.
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: integer
      - description: Sender ID
        in: path
        name: senderId
        required: true
        type: integer
      - description: Association Details
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_handler.AddSenderToAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/email-service_internal_model.AccountSender'
              type: object
        "400":
          description: Invalid ID or request body
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Failed to associate sender with account
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Add Sender to Account
      tags:
      - Senders
...
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and a JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"